Update examples README, rename files
